Moved app data to parent component to pass to the child components, Add button group for heatmap to zoom in, out, and return to normal scope, Add styles to leftDock, rightDock, footer, header, and content, update footer information and card_template.
