--- .vscode/settings.json
@@ -1,5 +1,0 @@
-{
-    "java.completion.enabled": false,
-    "java.debug.settings.enableRunDebugCodeLens": false,
-    "java.test.editor.enableShortcuts": false
-}
