--- nbproject/private/private.properties
@@ -1,1 +1,0 @@
-browser=SL[/Browsers/ChromeBrowser
