--- nbproject/project.properties
@@ -1,3 +1,0 @@
-files.encoding=UTF-8
-site.root.folder=
-source.folder=
