+++ package.json
@@ -1,17 +1,61 @@
 {
-  "name": "COVID19-Tracker",
-  "version": "1.0.0",
-  "scripts": {
-    "client-install": "yarn install --prefix client",
-    "start": "node server.js",
-    "heroku-postbuild": "YARN_CONFIG_PRODUCTION=false yarn install --cwd client && yarn --cwd client build"
-  },
-  "main": "index.js",
-  "repository": "https://github.com/mohur/COVID19-Tracker",
-  "author": "mohur <m.ashrahfi@gmail.com>",
-  "license": "MIT",
+  "name": "client",
+  "version": "0.1.0",
+  "private": true,
   "dependencies": {
-    "concurrently": "^5.3.0",
-    "express": "^4.17.1"
+    "@material-ui/core": "^4.11.0",
+    "@testing-library/jest-dom": "^4.2.4",
+    "@testing-library/react": "^9.3.2",
+    "@testing-library/user-event": "^7.1.2",
+    "axios": "^0.24.0",
+    "bootstrap": "^5.1.3",
+    "chart.js": "^2.9.3",
+    "classnames": "^2.2.6",
+    "d3-fetch": "^3.0.1",
+    "d3-geo": "^3.0.1",
+    "d3-scale": "^4.0.2",
+    "gh-pages": "^3.1.0",
+    "moment": "^2.29.1",
+    "react": "^16.13.1",
+    "react-bootstrap": "^2.0.0",
+    "react-chartjs-2": "^2.10.0",
+    "react-countup": "^4.3.3",
+    "react-dom": "^16.13.1",
+    "react-gauge-chart": "^0.4.0",
+    "react-heatmap-grid": "^0.8.2",
+    "react-icons": "^4.3.1",
+    "react-moment": "^0.9.7",
+    "react-scripts": "^3.4.3",
+    "react-simple-maps": "^2.3.0",
+    "react-spinners": "^0.11.0",
+    "react-strap": "^0.0.1",
+    "react-tooltip": "^4.2.21",
+    "react-transition-group": "^4.4.1",
+    "reactstrap": "^8.5.1",
+    "recharts": "^2.1.6",
+    "webpack": "^4.42.0"
+  },
+  "scripts": {
+    "predeploy": "yarn run build",
+    "deploy": "gh-pages -d build",
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
+    "eject": "react-scripts eject"
+  },
+  "eslintConfig": {
+    "extends": "react-app"
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
   }
 }
