--- server.js
@@ -1,16 +1,0 @@
-var express = require('express');
-var app = express();
-var path = require('path');
-
-if(process.env.NODE_ENV === 'production') {
-    //Set static folder
-    app.use(express.static('client/build'));
-
-    app.get('*', (req, res) => {
-        res.sendFile(path.resolve(__dirname, 'client', 'build', 'index.html'));
-    });
-}
-
-const port = process.env.PORT || 5000;
-
-app.listen(port, () => console.log(`Server has started on port ${port}`));
