+++ src/Components/Content/Charts/CapitaChart.js
