+++ src/Components/Footer/Footer.module.css
@@ -1,0 +1,41 @@
+.container {
+    display: flex;
+    align-items: center;
+    justify-items: space-around;
+    background-color: #52FFB8;
+    justify-content: center;
+    left: 0;
+    width: 100%;
+    bottom: 0;
+    position: fixed;
+  
+}
+
+.text_color {
+    color: black;
+}
+.logo {
+    width: 41px;
+    height:40px;
+    color: black;
+}
+
+.footerContent {
+    margin: 1em;
+    color: black;
+}
+
+a {
+    transition: background-position 275ms ease;
+    background: linear-gradient(to right, 	red, 	red 50%, black 50%);
+    background-clip: text;
+    -webkit-background-clip: text;
+    -webkit-text-fill-color: transparent;
+    background-size: 200% 100%;
+    background-position: 100%;
+    transition: background-position 275ms ease;
+  }
+  a:hover {
+    background-position: 0 100%;
+    color: black
+  }
