+++ src/Components/index.js
@@ -1,0 +1,6 @@
+export { default as Cards } from './Cards/Cards';
+export { default as Charts } from './Charts/Charts';
+export { default as CountryPicker } from './CountryPicker/CountryPicker';
+export { default as Title } from './Title/Title';
+export { default as Footer } from './Footer/Footer';
+export { default as LastUpdated } from './LastUpdated/LastUpdated';
